Use TreeComboBox for path combo boxes
Alternative to 'wx.lib.ComboTreeBox', which cannot be used within toolbars (inheriting from wrong class).
'TreeCtrlComboPopup' implements the 'wx.combo.ComboPopup' interface for 'wx.TreeCtrl'.

'TreeComboBox' is implemented as a 'wx.combo.ComboCtrl' using a 'TreeCtrlComboPopup'
and providing a public interace for the 'wx.TreeCtrl' inside 'TreeCtrlComboPopup'.

We store text and client data. Children are created in a different order from the tree view, which is ugly.