Add a TreeComboBox
As an alternative to ``wx.lib.ComboTreeBox`` which cannot be added to a
toolbar due to inheriting from the wrong parent class.
``TreeCtrlComboPopup`` implements the ``wx.combo.ComboPopup`` interface
for ``wx.TreeCtrl``.

``TreeComboBox`` is implemented as a ``wx.combo.ComboCtrl`` containing
the ``TreeCtrlComboPopup`` and providing a public interace for the
``wx.TreeCtrl`` inside the ``TreeCtrlComboPopup``.